+++ src/DocumentFormat.OpenXml/Validation/OpenXmlValidator.cs
@@ -4,6 +4,7 @@
 using DocumentFormat.OpenXml.Packaging;
 using System;
 using System.Collections.Generic;
+using System.Threading;
 
 namespace DocumentFormat.OpenXml.Validation
 {
@@ -17,10 +18,10 @@ namespace DocumentFormat.OpenXml.Validation
         private readonly DocumentValidator _documentValidator;
 
         /// <summary>
-        /// Initializes a new instance of the OpenXmlValidator.
+        /// Initializes a new instance of the <see cref="OpenXmlValidator"/>.
         /// </summary>
         /// <remarks>
-        /// Default to FileFormat.Office2007.
+        /// Defaults to <see cref="FileFormatVersions.Office2007"/>.
         /// </remarks>
         public OpenXmlValidator()
             : this(FileFormatVersions.Office2007)
@@ -28,12 +29,9 @@ namespace DocumentFormat.OpenXml.Validation
         }
 
         /// <summary>
-        /// Initializes a new instance of the OpenXmlValidator.
+        /// Initializes a new instance of the <see cref="OpenXmlValidator"/>.
         /// </summary>
         /// <param name="fileFormat">The target file format to be validated against.</param>
-        /// <remarks>
-        /// Default to FileFormat.Office2007.
-        /// </remarks>
         /// <exception cref="ArgumentOutOfRangeException">Thrown when the <paramref name="fileFormat"/> parameter is not a known format.</exception>
         public OpenXmlValidator(FileFormatVersions fileFormat)
         {
@@ -51,7 +49,7 @@ namespace DocumentFormat.OpenXml.Validation
 
         /// <summary>
         /// Gets or sets the maximum number of errors the OpenXmlValidator will return.
-        /// Default is 1000. A zero (0) value means no limitation.
+        /// Default is <c>1000</c>. A <c>0</c> value means no limitation.
         /// </summary>
         /// <exception cref="ArgumentOutOfRangeException">Throw when the value set is less than zero.</exception>
         public int MaxNumberOfErrors
@@ -72,11 +70,27 @@ namespace DocumentFormat.OpenXml.Validation
         /// <summary>
         /// Validates the specified document.
         /// </summary>
-        /// <param name="openXmlPackage">The target WordprocessingDocument, SpreadsheetDocument or PresentationDocument.</param>
+        /// <param name="openXmlPackage">The target <see cref="WordprocessingDocument"/>, <see cref="SpreadsheetDocument"/> or <see cref="PresentationDocument"/>.</param>
         /// <returns>A set of validation errors.</returns>
-        /// <exception cref="ArgumentNullException">Thrown when the "openXmlPackage" parameter is null.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlPackage"/> parameter is null.</exception>
         public IEnumerable<ValidationErrorInfo> Validate(OpenXmlPackage openXmlPackage)
+            => Validate(openXmlPackage, default);
+
+        /// <summary>
+        /// Validates the specified document.
+        /// </summary>
+        /// <param name="openXmlPackage">The target <see cref="WordprocessingDocument"/>, <see cref="SpreadsheetDocument"/> or <see cref="PresentationDocument"/>.</param>
+        /// <param name="token">Cancellation token</param>
+        /// <returns>A set of validation errors.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlPackage"/> parameter is null.</exception>
+#if FEATURE_CANCELLATION_TOKEN
+        public
+#else
+        private
+#endif
+        IEnumerable<ValidationErrorInfo> Validate(OpenXmlPackage openXmlPackage, CancellationToken token = default)
         {
+            // Note: This is private on .NET 3.5 as CancellationToken on that platform is not available.
             if (openXmlPackage == null)
             {
                 throw new ArgumentNullException(nameof(openXmlPackage));
@@ -93,18 +107,35 @@ namespace DocumentFormat.OpenXml.Validation
                 throw new InvalidOperationException(exceptionMessage);
             }
 
-            return _documentValidator.Validate(openXmlPackage, _settings);
+            return _documentValidator.Validate(openXmlPackage, _settings, token);
         }
 
         /// <summary>
-        /// Validates the specified content in the OpenXmlPart.
+        /// Validates the specified content in the <paramref name="openXmlPart"/>.
         /// </summary>
         /// <param name="openXmlPart">The target OpenXmlPart.</param>
         /// <returns>A set of validation errors.</returns>
-        /// <exception cref="ArgumentNullException">Thrown when the "openXmlPart" parameter is null.</exception>
-        /// <exception cref="InvalidOperationException">Throw when the specified part is not a defined part in the specified FileFormat version.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlPart"/> parameter is null.</exception>
+        /// <exception cref="InvalidOperationException">Throw when the specified part is not a defined part in the specified <see cref="FileFormat"/> version.</exception>
         public IEnumerable<ValidationErrorInfo> Validate(OpenXmlPart openXmlPart)
+            => Validate(openXmlPart, default);
+
+        /// <summary>
+        /// Validates the specified content in the <paramref name="openXmlPart"/>.
+        /// </summary>
+        /// <param name="openXmlPart">The target OpenXmlPart.</param>
+        /// <param name="token">Cancellation token</param>
+        /// <returns>A set of validation errors.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlPart"/> parameter is null.</exception>
+        /// <exception cref="InvalidOperationException">Throw when the specified part is not a defined part in the specified <see cref="FileFormat"/> version.</exception>
+#if FEATURE_CANCELLATION_TOKEN
+        public
+#else
+        private
+#endif
+        IEnumerable<ValidationErrorInfo> Validate(OpenXmlPart openXmlPart, CancellationToken token = default)
         {
+            // Note: This is private on .NET 3.5 as CancellationToken on that platform is not available.
             if (openXmlPart == null)
             {
                 throw new ArgumentNullException(nameof(openXmlPart));
@@ -124,7 +155,7 @@ namespace DocumentFormat.OpenXml.Validation
 
             FileFormat.ThrowIfNotInVersion(openXmlPart);
 
-            return _documentValidator.Validate(openXmlPart, _settings);
+            return _documentValidator.Validate(openXmlPart, _settings, token);
         }
 
         /// <summary>
@@ -132,11 +163,29 @@ namespace DocumentFormat.OpenXml.Validation
         /// </summary>
         /// <param name="openXmlElement">The target OpenXmlElement.</param>
         /// <returns>A set of validation errors.</returns>
-        /// <exception cref="ArgumentNullException">Thrown when the "openXmlElement" parameter is null.</exception>
-        /// <exception cref="ArgumentOutOfRangeException">Thrown when the "openXmlElement" is type of OpenXmlUnknownElement, OpenXmlMiscNode, AlternateContent, AlternateContentChoice or AlternateContentFallback.</exception>
-        /// <exception cref="InvalidOperationException">Thrown when the "openXmlElement" is not defined in the specified FileFormat.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlElement"/> parameter is null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when the <paramref name="openXmlElement"/> is type of <see cref="OpenXmlUnknownElement"/>, <see cref="OpenXmlMiscNode"/>, <see cref="AlternateContent"/>, <see cref="AlternateContentChoice"/> or <see cref="AlternateContentFallback"/>.</exception>
+        /// <exception cref="InvalidOperationException">Thrown when the <paramref name="openXmlElement"/> is not defined in the specified <see cref="FileFormat"/>.</exception>
         public IEnumerable<ValidationErrorInfo> Validate(OpenXmlElement openXmlElement)
+            => Validate(openXmlElement, default);
+
+        /// <summary>
+        /// Validates the specified element.
+        /// </summary>
+        /// <param name="openXmlElement">The target OpenXmlElement.</param>
+        /// <param name="token">Cancellation token</param>
+        /// <returns>A set of validation errors.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="openXmlElement"/> parameter is null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when the <paramref name="openXmlElement"/> is type of <see cref="OpenXmlUnknownElement"/>, <see cref="OpenXmlMiscNode"/>, <see cref="AlternateContent"/>, <see cref="AlternateContentChoice"/> or <see cref="AlternateContentFallback"/>.</exception>
+        /// <exception cref="InvalidOperationException">Thrown when the <paramref name="openXmlElement"/> is not defined in the specified <see cref="FileFormat"/>.</exception>
+#if FEATURE_CANCELLATION_TOKEN
+        public
+#else
+        private
+#endif
+        IEnumerable<ValidationErrorInfo> Validate(OpenXmlElement openXmlElement, CancellationToken token)
         {
+            // Note: This is private on .NET 3.5 as CancellationToken on that platform is not available.
             if (openXmlElement == null)
             {
                 throw new ArgumentNullException(nameof(openXmlElement));
@@ -161,7 +210,7 @@ namespace DocumentFormat.OpenXml.Validation
 
             FileFormat.ThrowIfNotInVersion(openXmlElement);
 
-            var validationContext = new ValidationContext(_settings, _cache);
+            var validationContext = new ValidationContext(_settings, _cache, token);
 
             using (validationContext.Stack.Push(element: openXmlElement))
             {